refactor: split transaction toolbar components
