+++ frontend/src/transactions/components/toolbar/TransactionListToolbar.tsx
@@ -1,0 +1,168 @@
+import { useCallback, useMemo, useState } from 'react'
+import type { TransactionListToolbarProps } from '@/transactions/components/toolbar/types'
+import { DesktopTransactionToolbarControls } from '@/transactions/components/toolbar/DesktopTransactionToolbarControls'
+import { MobileToolbarActions } from '@/transactions/components/toolbar/MobileToolbarActions'
+import { MobileTransactionFilterSheet } from '@/transactions/components/toolbar/MobileTransactionFilterSheet'
+import { TransactionSearchField } from '@/transactions/components/toolbar/TransactionSearchField'
+import { useDesktopToolbarLayout } from '@/transactions/components/toolbar/hooks/useDesktopToolbarLayout'
+import { useMobileSearchStuck } from '@/transactions/components/toolbar/hooks/useMobileSearchStuck'
+import { useToolbarStickyOffset } from '@/transactions/components/toolbar/hooks/useToolbarStickyOffset'
+import { formatDateRangeLabel } from '@/transactions/utils/date'
+import {
+  getAccountOptions,
+  getActiveFilterCount,
+  getCategoryOptions,
+} from '@/transactions/utils/filterOptions'
+
+/**
+ * Orchestrates transaction list search, filters, and responsive toolbar layout
+ */
+export default function TransactionListToolbar({
+  search,
+  onSearchChange,
+  onSearchSubmit,
+  filters,
+  setFilter,
+  categories,
+  accounts,
+  showAccountFilter,
+  pendingFrom,
+  pendingTo,
+  dateRangeChanged,
+  dateRangeInvalid,
+  onPendingFromChange,
+  onPendingToChange,
+  onDateRangeReset,
+  onDateRangeClose,
+  onCreateTransaction,
+  createDisabled = false,
+  createDisabledReason,
+  onStickyOffsetChange,
+}: TransactionListToolbarProps) {
+  const [isMobileSheetOpen, setIsMobileSheetOpen] = useState(false)
+  const [isMobileBackdropActive, setIsMobileBackdropActive] = useState(false)
+
+  const accountOptions = useMemo(
+    () => getAccountOptions(accounts),
+    [accounts],
+  )
+  const categoryOptions = useMemo(
+    () => getCategoryOptions(categories),
+    [categories],
+  )
+  const selectedAccountLabel = accounts?.find((account) => account.id === filters.account_id)?.name ?? null
+  const selectedCategoryLabel = categories?.find((category) => category.id === filters.category_id)?.name ?? null
+  const selectedDateLabel = formatDateRangeLabel(filters.from_date, filters.to_date)
+  const activeFilterCount = getActiveFilterCount(filters, showAccountFilter)
+
+  const {
+    toolbarRef,
+    controlsRef,
+    filterGroupRef,
+    createMeasureRef,
+    desktopInlineLayout,
+    desktopCreateStacked,
+  } = useDesktopToolbarLayout({
+    selectedAccountLabel,
+    selectedCategoryLabel,
+    selectedDateLabel,
+    showAccountFilter,
+  })
+  const { mobileSearchStickySentinelRef, mobileSearchStuck } = useMobileSearchStuck()
+
+  useToolbarStickyOffset(toolbarRef, onStickyOffsetChange)
+
+  const openMobileSheet = useCallback(() => {
+    setIsMobileBackdropActive(true)
+    setIsMobileSheetOpen(true)
+  }, [])
+
+  const closeMobileSheet = useCallback(() => {
+    onDateRangeClose()
+    setIsMobileSheetOpen(false)
+  }, [onDateRangeClose])
+
+  return (
+    <>
+      <div ref={mobileSearchStickySentinelRef} aria-hidden className="h-px min-[1050px]:hidden" />
+      <div
+        ref={toolbarRef}
+        className={`sticky top-0 z-30 !mt-2 mb-2 flex flex-col gap-3 pb-2 pt-4 min-[1050px]:pt-5 ${desktopInlineLayout ? 'min-[750px]:flex-row min-[750px]:items-center' : ''}`}
+        style={{
+          background: 'var(--app-bg)',
+          boxShadow: '0 0.25rem 0 var(--app-bg)',
+        }}
+      >
+        <TransactionSearchField
+          search={search}
+          onSearchChange={onSearchChange}
+          onSearchSubmit={onSearchSubmit}
+          mobileSearchStuck={mobileSearchStuck}
+          desktopInlineLayout={desktopInlineLayout}
+        />
+
+        <MobileToolbarActions
+          activeFilterCount={activeFilterCount}
+          onOpenFilters={openMobileSheet}
+          onCreateTransaction={onCreateTransaction}
+          createDisabled={createDisabled}
+          createDisabledReason={createDisabledReason}
+        />
+
+        <DesktopTransactionToolbarControls
+          filters={filters}
+          setFilter={setFilter}
+          showAccountFilter={showAccountFilter}
+          accountOptions={accountOptions}
+          categoryOptions={categoryOptions}
+          selectedAccountLabel={selectedAccountLabel}
+          selectedCategoryLabel={selectedCategoryLabel}
+          selectedDateLabel={selectedDateLabel}
+          pendingFrom={pendingFrom}
+          pendingTo={pendingTo}
+          dateRangeChanged={dateRangeChanged}
+          dateRangeInvalid={dateRangeInvalid}
+          onPendingFromChange={onPendingFromChange}
+          onPendingToChange={onPendingToChange}
+          onDateRangeReset={onDateRangeReset}
+          onDateRangeClose={onDateRangeClose}
+          desktopInlineLayout={desktopInlineLayout}
+          desktopCreateStacked={desktopCreateStacked}
+          controlsRef={controlsRef}
+          filterGroupRef={filterGroupRef}
+          createMeasureRef={createMeasureRef}
+          onCreateTransaction={onCreateTransaction}
+          createDisabled={createDisabled}
+          createDisabledReason={createDisabledReason}
+        />
+      </div>
+
+      {isMobileBackdropActive && (
+        <MobileTransactionFilterSheet
+          isOpen={isMobileSheetOpen}
+          activeFilterCount={activeFilterCount}
+          filters={filters}
+          setFilter={setFilter}
+          showAccountFilter={showAccountFilter}
+          accountOptions={accountOptions}
+          categoryOptions={categoryOptions}
+          selectedAccountLabel={selectedAccountLabel}
+          selectedCategoryLabel={selectedCategoryLabel}
+          selectedDateLabel={selectedDateLabel}
+          pendingFrom={pendingFrom}
+          pendingTo={pendingTo}
+          dateRangeChanged={dateRangeChanged}
+          dateRangeInvalid={dateRangeInvalid}
+          onPendingFromChange={onPendingFromChange}
+          onPendingToChange={onPendingToChange}
+          onDateRangeReset={onDateRangeReset}
+          onDateRangeClose={onDateRangeClose}
+          onClose={closeMobileSheet}
+          onExitComplete={() => {
+            if (!isMobileSheetOpen) setIsMobileBackdropActive(false)
+          }}
+        />
+      )}
+    </>
+  )
+}
