+++ frontend/src/transactions/components/toolbar/types.ts
@@ -1,0 +1,50 @@
+import type { RefObject } from 'react'
+import type { Category } from '@/api/categories'
+import type { OptionItem } from '@/components/FilterOptionList'
+import type { TransactionListAccount, TransactionListFilters } from '@/transactions/types/transactionList'
+
+export type TransactionFilterSetter = (patch: Partial<TransactionListFilters>) => void
+
+export type TransactionDateRangeDraftProps = {
+  pendingFrom: string
+  pendingTo: string
+  dateRangeChanged: boolean
+  dateRangeInvalid: boolean
+  onPendingFromChange: (value: string) => void
+  onPendingToChange: (value: string) => void
+  onDateRangeReset: () => void
+  onDateRangeClose: () => void
+}
+
+export type TransactionListToolbarProps = TransactionDateRangeDraftProps & {
+  search: string
+  onSearchChange: (value: string) => void
+  onSearchSubmit: () => void
+  filters: TransactionListFilters
+  setFilter: TransactionFilterSetter
+  categories?: Category[]
+  accounts?: TransactionListAccount[]
+  showAccountFilter: boolean
+  onCreateTransaction: () => void
+  createDisabled?: boolean
+  createDisabledReason?: string
+  onStickyOffsetChange?: (offset: number) => void
+}
+
+export type TransactionToolbarOptions = {
+  accountOptions: OptionItem[]
+  categoryOptions: OptionItem[]
+}
+
+export type TransactionToolbarSelectionLabels = {
+  selectedAccountLabel: string | null
+  selectedCategoryLabel: string | null
+  selectedDateLabel: string | null
+}
+
+export type DesktopTransactionToolbarRefs = {
+  toolbarRef: RefObject<HTMLDivElement | null>
+  controlsRef: RefObject<HTMLDivElement | null>
+  filterGroupRef: RefObject<HTMLDivElement | null>
+  createMeasureRef: RefObject<HTMLButtonElement | null>
+}
